File moved to Logging -> Log .hpp/cpp
delete MessageType.hpp, Severity.hpp
