+++ src/shared/Logging/Log.cpp
@@ -4,7 +4,7 @@ This file is released under the MIT license. See README-MIT for more information
 */
 
 #include "Log.hpp"
-#include "Logging/Logger.hpp"
+#include "Logger.hpp"
 #include "Utilities/Util.hpp"
 #include "Config/Config.h"
