+++ src/shared/Logging/Log.hpp
@@ -6,7 +6,7 @@ This file is released under the MIT license. See README-MIT for more information
 #pragma once
 
 #include "Common.hpp"
-#include "Logging/LoggerDefines.hpp"
+#include "LoggerDefines.hpp"
 #include "Config/Config.h"
 
 class WorldPacket;
