--- src/shared/Logging/MessageType.hpp
@@ -1,17 +1,0 @@
-/*
-Copyright (c) 2014-2023 AscEmu Team <http://www.ascemu.org>
-This file is released under the MIT license. See README-MIT for more information.
-*/
-
-#pragma once
-
-namespace AscEmu::Logging
-{
-    enum MessageType
-    {
-        TRACE,
-        DEBUG,
-        MINOR,
-        MAJOR
-    };
-}
