+++ src/world/Server/BroadcastMgr.cpp
@@ -12,7 +12,7 @@ This file is released under the MIT license. See README-MIT for more information
 #include "Storage/MySQLDataStore.hpp"
 #include "Storage/MySQLStructures.h"
 
-#include "Log.hpp"
+#include "Logging/Log.hpp"
 
 using AscEmu::Threading::AEThread;
 using std::chrono::milliseconds;
